+++ rocketgram/api/edit_message_live_location.py
@@ -22,4 +22,7 @@ class EditMessageLiveLocation(MessageOrBoolResultMixin, Request):
     chat_id: Optional[Union[int, str]] = None
     message_id: Optional[int] = None
     inline_message_id: Optional[str] = None
+    horizontal_accuracy: Optional[float] = None
+    heading: Optional[int] = None
+    proximity_alert_radius: Optional[int] = None
     reply_markup: Optional[INLINE_KEYBOARDS] = None