+++ rocketgram/api/inline_query_result_location.py
@@ -24,7 +24,10 @@ class InlineQueryResultLocation(InlineQueryResult):
     latitude: float
     longitude: float
     title: str
+    horizontal_accuracy: Optional[float] = None
     live_period: Optional[int] = None
+    heading: Optional[int] = None
+    proximity_alert_radius: Optional[int] = None
     reply_markup: Optional[InlineKeyboardMarkup] = None
     input_message_content: Optional[InputMessageContent] = None
     thumb_url: Optional[str] = None