+++ rocketgram/api/input_location_message_content.py
@@ -18,4 +18,7 @@ class InputLocationMessageContent(InputMessageContent):
 
     latitude: float
     longitude: float
+    horizontal_accuracy: Optional[float] = None
     live_period: Optional[int] = None
+    heading: Optional[int] = None
+    proximity_alert_radius: Optional[int] = None